% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_reml.R
\name{bin_reml}
\alias{bin_reml}
\title{Evaluate the maximum log-likelihood.}
\usage{
bin_reml(Y = NULL, CV = NULL, GK = NULL)
}
\arguments{
\item{Y}{A numeric matrix with one column containing phenotypic values.}

\item{CV}{A numeric matrix containing any user-specified covariates.}

\item{GK}{A numeric matrix containing marker scores.}
}
\value{
A list containing the following components: \describe{
  \item{\code{LL}} A numeric scalar. The -2*LL estimated by REML.
  \item{\code{vg}} A numeric scalar. The estimated genetic variance.
  \item{\code{ve}} A numeric scalar. The estimated error variance.}
}
\description{
Estimate genetic and error variance components for a phenotype using a set of
markers as explanatory variables.
}
\author{
Xiaolei Liu

Jiabo Wang

Qishan Wang

Feng Tian

Zhiwu Zhang

Aaron Kusmec
}
