% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmcpu.R
\name{farmcpu}
\alias{farmcpu}
\title{Perform GWAS using the FarmCPU model.}
\usage{
farmcpu(Y, GD, GM, CV = NULL, GP = NULL, method.sub = "reward",
  method.sub.final = "reward", method.bin = "static", bin.size = c(5e+05,
  5e+06, 5e+07), bin.selection = seq(10, 100, 10), memo = NULL,
  Prior = NULL, ncores.glm = 1, maxLoop = 10, converge = 1,
  iteration.output = FALSE, p.threshold = NA, ncores.reml = 1,
  threshold = 0.7)
}
\arguments{
\item{Y}{A dataframe with two columns: a character column of sample names and
a numeric column of phenotypic values. This dataframe may contain missing
values.}

\item{GD}{A \code{big.matrix} object.}

\item{GM}{A dataframe with three columns: a character column of sample names,
a numeric column of chromosomes, and a numeric column of base-pair
positions.}

\item{CV}{An optional numeric matrix of covariates.}

\item{GP}{A numeric vector of p-values to select pseudo-QTNs before
single-marker regression in the first iteration.}

\item{method.sub}{Character string indicating the method used to substitute
p-values for pseudo-QTNs. One of \code{c('reward', 'mean', 'median',
'penalty', 'onsite')}.}

\item{method.sub.final}{Character string indicating the method used to
substitute p-values for pseudo-QTNs on the final iteration.}

\item{method.bin}{Character string indicating the method used to select
pseudo-QTNs. One of \code{c('static', 'optimum')}.}

\item{bin.size}{A numeric vector specifying bin sizes in base-pairs for
pseudo-QTN selection.}

\item{bin.selection}{Numeric vector of numbers of pseudo-QTNs to select.}

\item{memo}{Character string added to file names.}

\item{Prior}{A dataframe with the same format as \code{GM} with a fourth
numeric column containing prior marker probabilities.}

\item{ncores.glm}{Integer scalar number of cores to use for single-marker
regression.}

\item{maxLoop}{Integer scalar maximum number of iterations to perform.}

\item{converge}{Numeric scalar such that \code{0 < converge < 1}. Controls
the percentage of overlapping pseudo-QTNs between two iterations required
to terminate the GWAS.}

\item{iteration.output}{Logical scalar. Whether to include the output of each
iteration or not.}

\item{p.threshold}{Numeric scalar. P-value threshold for inclusion of
pseudo-QTNs in the model on the first iteration. Defaults to a 0.01
Bonferroni correction.}

\item{ncores.reml}{Integer scalar number of cores to use for pseudo-QTN selection.}

\item{threshold}{Numeric scalar maximum correlation allowed between
pseudo-QTNs.}
}
\value{
A list. If \code{iteration.output = FALSE}, the list has one element.
  Otherwise, the list additionally has one element for each iteration. Each
  element has one required element and one optional element. The required
  element is a dataframe called \code{GWAS} with seven columns: the marker
  ID, the chromosome, the base-pair position, the marker p-value, the effect
  estimate, the standard error, and the t-statistic. If there are
  user-specified covariates, the results list contains a matrix called
  \code{CV} that contains the effect estimates for the covariates.
}
\description{
Perform GWAS using the FarmCPU model.
}
\author{
Xiaolei Liu

Zhiwu Zhang

Aaron Kusmec
}
