% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bins.R
\name{get_bins}
\alias{get_bins}
\title{Bin markers and select representative markers based on p-values.}
\usage{
get_bins(GI = NULL, GP = NULL, bin.size = 1e+07, nqtn = NULL,
  MaxBP = 1e+10)
}
\arguments{
\item{GI}{A data frame with three columns: marker name, chromosome, and
base-pair position.}

\item{GP}{A data frame with four columns: marker name, chromosome, base-pair
position, and p-value.}

\item{bin.size}{An integer scalar. The size (in base-pairs) of each bin.}

\item{nqtn}{An integer scalar. The number of pseudo-QTNs to select.}

\item{MaxBP}{An integer scalar. A scaling factor used to generate unique
marker IDs.}
}
\value{
A logical vector indicating whether or not a marker is a pseudo-QTN.
}
\description{
Create bins of an arbitrary size and select an arbitrary number of
representative markers across bins with the smallest p-values.
}
\details{
Markers are binned into bins of arbitrary size and represented by
  the marker with the smallest p-value within each bin.
}
\author{
Zhiwu Zhang
}

