% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qtns.R
\name{get_qtns}
\alias{get_qtns}
\title{Select pseudo-QTNs.}
\usage{
get_qtns(Y, GM, P, method, bin.sizes, nqtn, CV = NULL, GDP = NULL,
  theLoop = NULL)
}
\arguments{
\item{Y}{A data frame with two columns. The first column contains taxa names.
The second column contains numeric phenotypic values.}

\item{GM}{A data frame with three columns: marker IDs, chromosome, and
base-pair positions.}

\item{P}{A numeric vector of marker p-values.}

\item{method}{A character string, one of \code{c("static", "optimum")}.}

\item{bin.sizes}{A vector of possible bin sizes in base-pairs.}

\item{nqtn}{A vector of possible numbers of pseudo-QTNs to select.}

\item{CV}{A numeric matrix of user-specified covariates.}

\item{GDP}{A pointer to a big.matrix object containing marker scores.}

\item{theLoop}{The current iteration of the overal FarmCPU process. Controls
bin size selection for \code{method = "static"}.}
}
\value{
A logical vector specifying which markers have been chosen as
  pseudo-QTNs.
}
\description{
Use p-values estimated by \code{\link{quick_lm}} to select pseudo-QTNs for
use as covariates in subsequent single-marker regression tests. Pseudo-QTNs
may be selected using a set number of bins and a set number of pseudo-QTNs or
may be optimized based on a set of optional sizes.
}
\author{
Zhiwu Zhang

Aaron Kusmec
}
