% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Make a manhattan plot from GWAS results.}
\usage{
manhattan_plot(res, colors = c("black", "grey80"), cutoff = NULL,
  ylab = expression(paste(-log[10], "p-value")), linecol = "red",
  linetype = 2, linewidth = 2, ...)
}
\arguments{
\item{res}{A dataframe with at least three named columns: a numeric column of
chromosomes (\code{Chromosome}), a numeric column of base-pair positions
(\code{Position}), and a numeric column of marker p-values
(\code{p.value}).}

\item{colors}{A character vector of length two where elements are R color
names.}

\item{cutoff}{A significance cutoff (default = \code{NULL}).}

\item{ylab}{A label for the y-axis (default =
\code{expression(paste(-log[10], "p-value"))}).}

\item{linecol}{String specifying an R color for the cutoff line.}

\item{linetype}{Integer specifying the line type for the cutoff line.}

\item{linewidth}{Integer specifying the line width for the cutoff line.}

\item{...}{Other graphics arguments.}
}
\description{
Make a manhattan plot from GWAS results.
}
\author{
Aaron Kusmec
}
