% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\title{Set marker p-values based on prior expectations.}
\usage{
prior(GM, P = NULL, Prior = NULL)
}
\arguments{
\item{GM}{A dataframe with three columns containing the SNP ID, chromosome,
and base-pair position of each marker.}

\item{P}{A numeric vector containing marker p-values.}

\item{Prior}{A dataframe with four columns. The first three correspond to the
columns of \code{GM} and the fourth contains prior probabilities.}
}
\value{
A numeric vector of updated marker p-values.
}
\description{
Modifies marker p-values calculated by the GLM based on prior expectations.
}
\details{
If \code{Prior} is \code{NULL}, p-values are not modified. Marker
  p-values are multiplied by the user-provided prior p-values. A value of 1
  will not modify the p-values calculated by the GLM.
}
\author{
Zhiwu Zhang

Aaron Kusmec
}

