% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_lm.R
\name{quick_lm}
\alias{quick_lm}
\title{Perform single marker regression tests.}
\usage{
quick_lm(Y = NULL, GDP = NULL, CV = NULL, npc = 0, seqQTN = NULL)
}
\arguments{
\item{Y}{A numeric vector of phenotype values. May include missing values.}

\item{GDP}{A pointer to a big.matrix object.}

\item{CV}{A numeric matrix where each column contains values for a covariate.}

\item{npc}{An integer scalar. Number of columns of \code{CV} that contain
user-defined covariates.}

\item{seqQTN}{An integer vector denoting the column in \code{GDP} of each
pseudo-QTN.}
}
\value{
A list containing the following components:
  \describe{
    \item{\code{P}}{a numeric matrix whose first four columns contain
  the effect estimate, standard error, t statistic, and p-value for each
  single marker regression test. Subsequent columns contain p-values for
  pseudo-QTNs when used as a covariate when testing each
  marker.}
    \item{\code{P0}}{a numeric vector containing p-values for the mean,
  user-specified covariates, and pseudo-QTNs in the null
  model.}
    \item{\code{betapc}}{a numeric vector of effect estimates for
  user-specified covariates.}}
}
\description{
Returns single marker regression test results for all markers in \code{GDP},
effect estimates for covariates (user-specified and pseudo-QTNs), and
p-values for pseudo-QTNs used by \code{\link{get_qtns}}.
}
\author{
Aaron Kusmec
}

