% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_qtns.R
\name{remove_qtns}
\alias{remove_qtns}
\title{Remove highly correlated pseudo-QTNs.}
\usage{
remove_qtns(GDP = NULL, GM = NULL, seqQTN = NULL, seqQTN.p = NULL,
  threshold = 0.99)
}
\arguments{
\item{GDP}{A pointer to a big.matrix object containing the marker scores.}

\item{GM}{A data frame with three columns: marker ID, chromosome, and
base-pair position.}

\item{seqQTN}{An integer vector of columns in GDP that have been selected as
pseudo-QTNs.}

\item{seqQTN.p}{A numeric vector of p-values for the pseudo-QTNs.}

\item{threshold}{A numeric scalar. Pairs of pseudo-QTNs with correlation
coefficients above this value will be pruned.}
}
\value{
An integer vector specifying the columns in \code{GDP} that are
  selected as pseudo-QTNs.
}
\description{
Removes highly correlated pseudo-QTNs based on the Pearson correlation
coefficients between marker scores. In each pair of highly correlated
markers, the marker with the lower p-value is retained.
}
\author{
Zhiwu Zhang

Aaron Kusmec
}
