% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_pvalue.R
\name{substitute_pvalue}
\alias{substitute_pvalue}
\title{Substitute p-values for pseudo-QTNs.}
\usage{
substitute_pvalue(GM, GLM, QTN = NULL, method = "mean", npc = 0)
}
\arguments{
\item{GM}{A dataframe with three columns containing the SNP ID, chromosome,
and base-pair position of each marker.}

\item{GLM}{A \code{FarmCPUpp} object returned by \code{\link{quick_lm}}.}

\item{QTN}{A three column matrix containing the SNP ID, chromosome, and
base-pair position of each pseudo-QTN.}

\item{method}{A character scalar, one of "penalty", "reward", "mean",
"median", or "onsite".}

\item{npc}{An integer scalar. The number of user-defined covariates.}
}
\value{
The input \code{GLM} object with the pseudo-QTN p-values in
  \code{GLM[["P"]]} substituted.
}
\description{
Replace the p-values calculated for pseudo-QTNs according to one of several
rules.
}
\details{
Fill in details.
}
\author{
Xiaolei Liu

Zhiwu Zhang

Aaron Kusmec
}
