% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.R
\name{write_results}
\alias{write_results}
\title{Write GWAS results to disk.}
\usage{
write_results(res)
}
\arguments{
\item{res}{A results list returned by \code{\link{farmcpu}}.}
}
\description{
Write GWAS results to disk.
}
\author{
Aaron Kusmec
}
